Refactor init module and trust handling
Closes #147
